+++ complexity/batch.sh
@@ -11,35 +11,35 @@ TMPFILE="${SCRATCHDIR}/tmp/tmp-$2"
 
 DATAFILE="$(Rscript gen_data.R $BATCHSET)"
 
-if [ "$?" == "0" ]; then
+if [ "$?" == "0" ] && [ -f "$DATAFILE" ]; then
 
 	rm -f $TMPFILE
 
 	# Current version of optmatch doesn't work with Rscript.
 	# For future versions, use Rscript if possible. I.e.:
-	# `./time.sh Rscript match_funcs/scclust_EXU_CSE.R $DATAFILE >> $TMPFILE`
+	# `./time.sh Rscript do_matching.R scclust_EXU_CSE $DATAFILE >> $TMPFILE`
 
 	if [ "$TORUN" == "level1" ]; then
-		./time.sh R --vanilla --slave "--args $DATAFILE" < match_funcs/opt_kmatch.R >> $TMPFILE
+		./time.sh R --vanilla --slave "--args opt_kmatch $DATAFILE" < do_matching.R >> $TMPFILE
 	fi
 
 	if [ "$TORUN" == "level1" ] || [ "$TORUN" == "level2" ]; then
-		./time.sh R --vanilla --slave "--args $DATAFILE" < match_funcs/opt_pairmatch.R >> $TMPFILE
+		./time.sh R --vanilla --slave "--args opt_pairmatch $DATAFILE" < do_matching.R >> $TMPFILE
 	fi
 
 	if [ "$TORUN" == "level1" ] || [ "$TORUN" == "level2" ] || [ "$TORUN" == "level3" ]; then
-		./time.sh R --vanilla --slave "--args $DATAFILE" < match_funcs/opt_fullmatch.R >> $TMPFILE
+		./time.sh R --vanilla --slave "--args opt_fullmatch $DATAFILE" < do_matching.R >> $TMPFILE
 	fi
 
 	if [ "$TORUN" == "level1" ] || [ "$TORUN" == "level2" ] || [ "$TORUN" == "level3" ] || [ "$TORUN" == "level4" ]; then
-		./time.sh R --vanilla --slave "--args $DATAFILE" < match_funcs/gre_kmatch.R >> $TMPFILE
-		./time.sh R --vanilla --slave "--args $DATAFILE" < match_funcs/gre_pairmatch.R >> $TMPFILE
-		./time.sh R --vanilla --slave "--args $DATAFILE" < match_funcs/rep_pairmatch.R >> $TMPFILE
+		./time.sh R --vanilla --slave "--args gre_kmatch $DATAFILE" < do_matching.R >> $TMPFILE
+		./time.sh R --vanilla --slave "--args gre_pairmatch $DATAFILE" < do_matching.R >> $TMPFILE
+		./time.sh R --vanilla --slave "--args rep_pairmatch $DATAFILE" < do_matching.R >> $TMPFILE
 	fi
 
 	if [ "$TORUN" == "level1" ] || [ "$TORUN" == "level2" ] || [ "$TORUN" == "level3" ] || [ "$TORUN" == "level4" ] || [ "$TORUN" == "level5" ]; then
-		./time.sh R --vanilla --slave "--args $DATAFILE" < match_funcs/scclust_EXU_CSE.R >> $TMPFILE
-		./time.sh R --vanilla --slave "--args $DATAFILE" < match_funcs/scclust_LEX_ANY.R >> $TMPFILE
+		./time.sh R --vanilla --slave "--args scclust_EXU_CSE $DATAFILE" < do_matching.R >> $TMPFILE
+		./time.sh R --vanilla --slave "--args scclust_LEX_ANY $DATAFILE" < do_matching.R >> $TMPFILE
 	fi
 
 	cat $TMPFILE >> $OUTFILE
